+++ awc/tests/test_ws.rs
@@ -4,8 +4,8 @@ use actix_codec::Framed;
 use actix_http::{body::BodySize, h1, ws, Error, HttpService, Request, Response};
 use actix_http_test::test_server;
 use bytes::Bytes;
-use futures::future::ok;
-use futures::{SinkExt, StreamExt};
+use futures_util::future::ok;
+use futures_util::{SinkExt, StreamExt};
 
 async fn ws_service(req: ws::Frame) -> Result<ws::Message, io::Error> {
     match req {