+++ src/responder.rs
@@ -10,8 +10,8 @@ use actix_http::http::{
 };
 use actix_http::{Error, Response, ResponseBuilder};
 use bytes::{Bytes, BytesMut};
-use futures::future::{err, ok, Either as EitherFuture, Ready};
-use futures::ready;
+use futures_util::future::{err, ok, Either as EitherFuture, Ready};
+use futures_util::ready;
 use pin_project::{pin_project, project};
 
 use crate::request::HttpRequest;