+++ actix-http/CHANGES.md
@@ -7,6 +7,9 @@
 * Bump minimum supported Rust version to 1.40
 * content_length function is removed, and you can set Content-Length by calling no_chunking function [#1439]
 
+* `BodySize::Sized64` variant has been removed. `BodySize::Sized` now receives a
+  `u64` instead of a `usize`.
+
 ### Fixed
 
 * Support parsing of `SameSite=None` [#1503]