+++ actix-http/CHANGES.md
@@ -6,9 +6,9 @@
 
 * Bump minimum supported Rust version to 1.40
 * content_length function is removed, and you can set Content-Length by calling no_chunking function [#1439]
-
 * `BodySize::Sized64` variant has been removed. `BodySize::Sized` now receives a
   `u64` instead of a `usize`.
+* Update `base64` dependency to 0.12
 
 ### Fixed