+++ actix-http/CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-## [Unreleased]
+## [2.0.0-alpha.4] - 2020-05-21
 
 ### Changed